+++ Source/Day21.cpp
@@ -38,14 +38,14 @@ void rotate(string& s, bool left, int count)
     {
         if(left)
         {
-            for(int i = 0; i < s.length()-1; ++i)
+            for(int i = 0; i < static_cast<int>(s.length())-1; ++i)
             {
                 swap(s, i, i+1);
             }
         }
         else
         {
-            for(int i = int(s.length())-1; i > 0; --i)
+            for(int i = static_cast<int>(s.length())-1; i > 0; --i)
             {
                 swap(s, i, i+1);
             }
